requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.8
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
entrypoints==0.3
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
Pygments==2.6.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.15.0.1.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.4
webencodings==0.5.1
zipp==3.1.0
